% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eg.R
\name{eg}
\alias{eg}
\title{eg}
\usage{
eg(object)
}
\arguments{
\item{object}{a matrix or vector}
}
\value{
an output
}
\description{
eg gives description of the dataset
}
\examples{
x <- matrix(c(1,2,3,4,5,6), nrow = 2)
summary(x)
}
