% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flagged.R
\docType{data}
\name{Flagged}
\alias{Flagged}
\title{Flagged vector}
\format{
A matrix with 100 rows and 1 variables:
\describe{
\item{Flagged}{variable to indicate whether the test vendor suspects the examinee may have engaged in inappropriate behavior. 1= cheating cases, 0= non-cheating cases}
}
}
\usage{
Flagged
}
\description{
A vector containing information about 100 candidates whether they are flagged or not. A flagged candidate is believed to have engaged in test
fraud or cheating.
}
\details{
0 represent cheating cases
1 represent non-cheating cases
}
\keyword{datasets}
