% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histall.R
\name{histall}
\alias{histall}
\title{histall}
\usage{
histall(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
histogram
}
\description{
histall gives histogram of all numeric variables
}
\examples{
x <- as.data.frame(matrix(c(1,2,3,4,5,6), nrow = 2))
histall(x)
}
