% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.SecMin.R
\name{plot.SecMin}
\alias{plot.SecMin}
\title{plot.SecMin}
\usage{
\method{plot}{SecMin}(
  x,
  alpha = NULL,
  beta = NULL,
  speed = NULL,
  plot.type = c("trace", "histogram", "autocorr", "ecdf"),
  ...
)
}
\arguments{
\item{x}{objects from class \code{SecMin}}

\item{alpha}{item numbers of the alpha parameter interested}

\item{beta}{item numbers of the alpha parameter interested}

\item{speed}{person IDs of the speed parameter interested}

\item{plot.type}{plot types}

\item{...}{graphical parameters of \code{plot()}}
}
\value{
trace, histrogram, autocorrelation and ECDF plots of alpha, beta, and speed parameters
}
\description{
plot.SecMin fit the lognormal model and return parameter estimates
}
