% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare}
\alias{prepare}
\title{prepare}
\usage{
prepare(RT, Resp)
}
\arguments{
\item{RT}{data frame containing RT information}

\item{Resp}{data frame containing item response}
}
\description{
prepare prepare the datasets and create values in the Global Environment to be used in secmin later
}
