% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepftr.R
\name{prepftr}
\alias{prepftr}
\title{prepftr}
\usage{
prepftr(SecMin.obj, RT, Resp, Flgd, Info = NULL)
}
\arguments{
\item{SecMin.obj}{objects from class \code{SecMin}}

\item{RT}{A required \eqn{N × J} \code{matrix} containing the response
times of \eqn{N} individuals to \eqn{J} items. Zero and missing values need
to be coded as \code{NA}.}

\item{Resp}{A required \eqn{N × J} \code{matrix} or \code{data.frame}
containing the item responses of \eqn{N} individuals to \eqn{J} items. The
correct responses are marked by 1 and incorrect responses by 0. Missing values
need to be coded as \code{NA}.}

\item{Flgd}{A required vector containing whether the individual is marked as
cheating or not. Non-cheaters need to be marked by 1 and cheaters need to be
marked by 0.}

\item{Info}{A non-required \eqn{N × I} \code{matrix} or \code{data.frame}
containing the background information of \eqn{N} individuals to \eqn{I} variables.}
}
\value{
\code{prepftr} returns the name and variable number of highly correlated
variables and saves two datasets, \code{full.features} and \code{selected.features}.
\code{full.features} include LZ values, Ht values,
Lzstart values, theta values from IRT analysis, person speed estimates, item
responses, and RTs. \code{selected.features} are \code{full.features} excluding
the highly correlated variables.
}
\description{
prepftr prepares the features to be used in supervised and unsupervised machine
learning models
}
