% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepini.R
\name{prepini}
\alias{prepini}
\title{prepini}
\usage{
prepini(RT, Resp)
}
\arguments{
\item{RT}{A required \eqn{N × J} \code{matrix} or \code{data.frame}
containing the response times of \eqn{N} individuals to \eqn{J} items. Missing
values need to be coded as \code{NA}.}

\item{Resp}{A required \eqn{N × J} \code{matrix} or \code{data.frame}
containing the item responses of \eqn{N} individuals to \eqn{J} items. The
correct responses are marked by 1 and incorrect responses by 0. Missing values
need to be coded as \code{NA}.}
}
\value{
\code{prepini} returns two datasets and several initial values
\item{RT}{The RT data is saved as a matrix with zero values marked as \code{NA}.}
\item{Resp}{The Resp data is saved as a \code{data.frame}.}
\item{ini.alpha}{A vector containing initial alpha values for each item.}
\item{ini.beta}{A vector containing initial beta values for each item.}
\item{ini.speed}{A vector containing initial person speed values for each individual.}
\item{ini.prec.speed}{The initial value for the precision of speed}
\item{ini.mu.ipar}{The initial values for the means of the item parameters.}
\item{ini.prec.ipar}{The initial values for the variance-covariance matrix of the item parameters.}
}
\description{
prepini prepares the initial values to be used in the lognormal model later
and saves the RT dataset as \code{matrix} and the Resp dataset as a \code{data.frame}
}
