% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.R
\name{prim}
\alias{prim}
\title{prim}
\usage{
prim(RT, Resp)
}
\arguments{
\item{RT}{data frame containing RT information}

\item{Resp}{data frame containing item response}
}
\value{
a list
\itemize{
\item out = the results of the lognormal model
\item alpha.table = estimates of alpha values
\item beta.table = estimates of beta values
\item speed.table = estimates of person speed
\item LZT.cheating.cases = predicted cheating cases based on LZT index
\item KLD.cheating.cases = predicted cheating cases based on KLD index
}
}
\description{
prim fit the lognormal model and return parameter estimates
}
