% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RT.R
\docType{data}
\name{RT}
\alias{RT}
\title{RT dataset}
\format{
A matrix with 1636 rows and 20 variables:
\describe{
\item{idur.1}{XXXX}
\item{idur.2}{XXXX}
\item{idur.3}{XXXX}
\item{idur.4}{XXXX}
\item{idur.5}{XXXX}
\item{idur.6}{XXXX}
\item{idur.7}{XXXX}
\item{idur.8}{XXXX}
\item{idur.9}{XXXX}
\item{idur.10}{XXXX}
\item{idur.11}{XXXX}
\item{idur.12}{XXXX}
\item{idur.13}{XXXX}
\item{idur.14}{XXXX}
\item{idur.15}{XXXX}
\item{idur.16}{XXXX}
\item{idur.17}{XXXX}
\item{idur.18}{XXXX}
\item{idur.19}{XXXX}
\item{idur.20}{XXXX}
}
}
\usage{
RT
}
\description{
A dataset containing response time of 1636 test-takers and 20 items
}
\keyword{datasets}
