% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secmin.R
\name{secmin}
\alias{secmin}
\title{secmin}
\usage{
secmin(RT, Resp)
}
\arguments{
\item{RT}{data frame containing RT information}

\item{Resp}{data frame containing item response}
}
\value{
a list
}
\description{
secmin fit the lognormal model and return parameter summaries and plots
}
