% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprvs.R
\name{sprvs}
\alias{sprvs}
\title{sprvs}
\usage{
sprvs(data, Flgd, ntree = 150)
}
\arguments{
\item{data}{A required \eqn{N × K} \code{matrix} or \code{data.frame}
containing \eqn{K} features of \eqn{N} individuals to be fitted for supervised
machine learning models. \code{full.features} and \code{selected.features} produced
by \code{\link{prepftr}} can be directly input here.}

\item{Flgd}{A dichotomous vector marking whether a particular participant is
marked as a cheater or not. Non-cheaters need to be marked by 1 and cheaters
need to be marked by 0.}

\item{ntree}{The number of trees to grow of the random forest algorithm. The default number is 150.}
}
\value{
\code{sprvs} returns the results from K-nearest neighbor (KNN) and random forest
\item{KNN Results}{reports the confusion matrix,and several performance metric
including accuracy, kappa, Mcnemar's Test p-value, sensitivity, specificity, positive
predictive value, negative predictive value, prevalence, detection rate, detection
prevalence, and the balanced accuracy from KNN}
\item{Random Forest Results}{reports the confusion matrix,and several performance metric
including accuracy, kappa, Mcnemar's Test p-value, sensitivity, specificity, positive
predictive value, negative predictive value, prevalence, detection rate, detection
prevalence, and the balanced accuracy by random forest}
}
\description{
sprvs fit gives cheating casess based on supervised machine learning models
}
