% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsprvs.R
\name{unsprvs}
\alias{unsprvs}
\title{unsprvs}
\usage{
unsprvs(data, Flgd)
}
\arguments{
\item{data}{A required \eqn{N × K} \code{matrix} or \code{data.frame}
containing \eqn{K} features of \eqn{N} individuals to be fitted for unsupervised
machine learning models. \code{full.features} and \code{selected.features} produced
by \code{\link{prepftr}} can be directly input here.}

\item{Flgd}{A dichotomous vector marking whether a particular participant is
marked as a cheater or not. Non-cheaters need to be marked by 1 and cheaters
need to be marked by 0.}
}
\value{
\code{unsprvs} returns the results from Kmeans and finite mixture
modeling (FMM)
\item{Detected Cheating Cases Using Kmeans}{reports the cheating cases detected
by Kmeans}
\item{Kmeas Results}{reports the confusion matrix,and several performance metric
including accuracy, kappa, Mcnemar's Test p-value, sensitivity, specificity, positive
predictive value, negative predictive value, prevalence, detection rate, detection
prevalence, and the balanced accuracy from Kmeans analysis}
\item{Detected Cheating Cases Using FMM}{reports
the cheating cases detected by FMM}
\item{FMM Results}{reports the confusion matrix,and several performance metric
including accuracy, kappa, Mcnemar's Test p-value, sensitivity, specificity, positive
predictive value, negative predictive value, prevalence, detection rate, detection
prevalence, and the balanced accuracy by FMM}
}
\description{
unsprvs gives cheating cases based on unsupervised machine learning models
}
