% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.R
\name{prim}
\alias{prim}
\title{prim}
\usage{
prim(RT, Resp)
}
\arguments{
\item{RT}{data frame containing RT information}

\item{Resp}{data frame containing item response}
}
\value{
a list
}
\description{
prim fit the lognormal model and return parameter estimates
}
