% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.R, R/secmin.R
\name{secmin}
\alias{secmin}
\title{prim}
\usage{
secmin(RT, Resp)

secmin(RT, Resp)
}
\arguments{
\item{RT}{data frame containing RT information}

\item{Resp}{data frame containing item response}
}
\value{
a list

a list
}
\description{
prim fit the lognormal model and return parameter estimates

secmin fit the lognormal model and return parameter summaries and plots
}
